requirements.txt
farm-haystack
streamlit
torch
pytube
git+https://github.com/openai/whisper.git 
llama-cpp-python
